Remove cache after change data for bug data not change after update

+++ src/External/Presentation/PenomyAPI.Presentation.FastEndpointBasedApi/Helpers/Cache/ICommonCacheHandler.cs
@@ -1,17 +1,16 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 using PenomyAPI.Presentation.FastEndpointBasedApi.Endpoints.FeatG5.Common;
 using PenomyAPI.Presentation.FastEndpointBasedApi.Endpoints.FeatG5.DTOs;
 using PenomyAPI.Presentation.FastEndpointBasedApi.Endpoints.FeatG5.HttpResponse;
 using PenomyAPI.Presentation.FastEndpointBasedApi.Endpoints.FeatG9.DTOs;
 using PenomyAPI.Presentation.FastEndpointBasedApi.Endpoints.FeatG9.HttpResponses;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace PenomyAPI.Presentation.FastEndpointBasedApi.Helpers.Cache;
 
 public interface ICommonCacheHandler
 {
-    ValueTask ClearG5MangaDetailCacheAsync(int mangaId, CancellationToken ct);
+    ValueTask ClearG5MangaDetailCacheAsync(long mangaId, CancellationToken ct);
 
     Task<G5HttpResponse> GetOrSetG5MangaDetailCacheAsync(
         G5StateBag stateBag,
